🤖 refactor(site/src/pages): nest health pages
