+++ site/src/pages/HealthPage/AccessURLPage/AccessURLPage.stories.tsx
@@ -2,14 +2,14 @@ import type { StoryObj } from "@storybook/react-vite";
 import { HEALTH_QUERY_KEY } from "#/api/queries/debug";
 import type { HealthcheckReport } from "#/api/typesGenerated";
 import { MockHealth } from "#/testHelpers/entities";
-import { generateMeta } from "./storybook";
-import WebsocketPage from "./WebsocketPage";
+import { generateMeta } from "../storybook";
+import AccessURLPage from "./AccessURLPage";
 
 const meta = {
-	title: "pages/Health/Websocket",
+	title: "pages/HealthPage/AccessURLPage",
 	...generateMeta({
-		path: "/health/websocket",
-		element: <WebsocketPage />,
+		path: "/health/access-url",
+		element: <AccessURLPage />,
 	}),
 };
 
@@ -21,8 +21,8 @@ const Example: Story = {};
 const settingsWithError: HealthcheckReport = {
 	...MockHealth,
 	severity: "error",
-	websocket: {
-		...MockHealth.websocket,
+	access_url: {
+		...MockHealth.access_url,
 		severity: "error",
 		error:
 			'EACS03: get healthz endpoint: Get "https://localhost:7080/healthz": http: server gave HTTP response to HTTPS client',
@@ -41,4 +41,4 @@ export const WithError: Story = {
 	},
 };
 
-export { Example as Websocket };
+export { Example as AccessURL };
