+++ site/src/pages/HealthPage/DERPRegionPage/DERPRegionPage.stories.tsx
@@ -1,13 +1,13 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { MockHealth } from "#/testHelpers/entities";
+import { generateMeta } from "../storybook";
 import DERPRegionPage from "./DERPRegionPage";
-import { generateMeta } from "./storybook";
 
 const firstRegionId = Object.values(MockHealth.derp.regions)[0]!.region
 	?.RegionID;
 
 const meta: Meta = {
-	title: "pages/Health/DERPRegion",
+	title: "pages/HealthPage/DERPRegionPage",
 	...generateMeta({
 		path: "/health/derp/regions/:regionId",
 		element: <DERPRegionPage />,
