+++ site/src/pages/HealthPage/DatabasePage/DatabasePage.tsx
@@ -11,8 +11,8 @@ import {
 	HealthMessageDocsLink,
 	HealthyDot,
 	Main,
-} from "./Content";
-import { MuteWarningsButton } from "./MuteWarningsButton";
+} from "../Content";
+import { MuteWarningsButton } from "../MuteWarningsButton";
 
 const DatabasePage = () => {
 	const healthStatus = useOutletContext<HealthcheckReport>();
