+++ site/src/pages/HealthPage/ProvisionerDaemonsPage/ProvisionerDaemonsPage.tsx
@@ -10,8 +10,8 @@ import {
 	HealthMessageDocsLink,
 	HealthyDot,
 	Main,
-} from "./Content";
-import { MuteWarningsButton } from "./MuteWarningsButton";
+} from "../Content";
+import { MuteWarningsButton } from "../MuteWarningsButton";
 
 const ProvisionerDaemonsPage: FC = () => {
 	const healthStatus = useOutletContext<HealthcheckReport>();
